Rework image searching and loading
Icons are now searched across all registered image search paths,
but will only be recolored if found in a path that was registered
with an icon style.

The actual loading of the image has been pushed down into the
path searching logic in order to allow for the search to continue
if the file exists but otherwise fails to load.

Also allow icon names to include extensions, so as not to violate
the Principal of Least Astonishment for doc-skimmers.