Fix modal release bug when touchscroll is enabled
When a modal widget (such as rtk.Popup) is set during a mousedown
handler, when touchscreen is enabled (in which the touch activate event
is mouse up) the subsequent mouseup event would cause the modal
widget(s) to immediately release.

This skips the release request if the modal widgets were set in the same
tick as the touch activate event occurred.